refactor(p2p): extract the P2P stats engine from the settings component
p2p-stats-settings.tsx was a 1210-line file that owned libp2p transfer-stat
instrumentation, peer normalization, public-endpoint resolution, polling, and
reducer state alongside its rendering. Split it into three layers:

- src/lib/p2p-stats.ts: the pure data-collection engine (transfer helpers,
  metric-snapshot parsing, peer normalization, endpoint resolution, getP2PStats)
- src/hooks/use-p2p-stats.ts: the polling lifecycle (reducer + abortable 5s
  refresh), returning { error, loading, rows, updatedAt }
- the component (now 176 lines): rendering only

The engine is a verbatim move and the polling semantics are unchanged, so the
rendered table, loading/error/starting states, and refresh behavior are
identical. Adds focused lib tests for byte formatting/merging, peer-record
normalization, transport-label detection, public/private IP filtering, and
no-throw handling of malformed metric snapshots.

Implements plans/004-extract-p2p-stats-engine.md.
